@@ -1,20 +1,19 @@
 name: qr_code_scanner
 description: QR code scanner that can be embedded inside flutter. It uses zxing in Android and MTBBarcode scanner in iOS.
-version: 0.0.13
+version: 0.0.13-nullsafety
 author: Julius Canute <juliuscanute at touchcapture.net>
 homepage: https://juliuscanute.com
 repository: https://github.com/juliuscanute/qr_code_scanner
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
-  flutter: ^1.10.0
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  pedantic: ^1.9.2
 
 flutter:
   plugin:
